Separate the async client from the EventSocket, clippy, add is_connected() function to the EventSocket
